Moved css functions to its own util file
