+++ client/pages/edit/events/[[...id]].tsx
@@ -4,7 +4,8 @@ import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { useForm } from 'react-hook-form';
 import Cookies from 'universal-cookie';
 import dayjs, { Dayjs } from 'dayjs';
-import { createPopupEvent, createEvent, darkSwitch, getTokenFromCookies } from '../../../src/util';
+import { createPopupEvent, createEvent, getTokenFromCookies } from '../../../src/util';
+import { darkSwitch } from "../../../src/util/cssUtil";
 import { getEvent, getOverlappingReservations, getUserInfo, postEvent, putEvent } from '../../../src/api';
 import { AccessLevel, EventType, PopupEvent } from '../../../src/types';
