+++ client/pages/edit/history/[resource]/[id].tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from 'next';
 import { useRouter } from 'next/router';
-import { calculateEditDate, darkSwitch, getParams } from '../../../../src/util';
+import { calculateEditDate, getParams } from '../../../../src/util';
+import { darkSwitch } from "../../../../src/util/cssUtil";
 import { getHistory } from '../../../../src/api';
 import type { History, Resource } from '../../../../src/types';
