+++ client/src/components/clubs/exec-card.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { Theme } from '@mui/material';
-import { darkSwitchGrey } from '../../util';
+import { darkSwitchGrey } from "../../util/cssUtil";
 import type { Exec } from '../../types';
 
 import Box from '@mui/material/Box';
