+++ client/src/components/edit/shared/edit-login.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import Cookies from 'universal-cookie';
 import type { PopupEvent } from '../../../types';
-import { createConnectionErrorPopup, darkSwitch, darkSwitchGrey } from '../../../util';
+import { createConnectionErrorPopup } from '../../../util';
+import { darkSwitch, darkSwitchGrey } from "../../../util/cssUtil";
 import { getUserInfo } from '../../../api';
 
 import Box from '@mui/material/Box';
