+++ client/src/components/reservations/reservation-day.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import type { Dayjs } from 'dayjs';
 import { Theme } from '@mui/material';
-import { darkSwitch } from '../../util';
+import { darkSwitch } from "../../util/cssUtil";
 import type { BrokenReservation } from '../../types';
 
 import Box from '@mui/material/Box';
