+++ client/src/components/reservations/reservation-month.tsx
@@ -1,7 +1,8 @@
 import React, { useEffect, useState } from 'react';
 import type { Dayjs } from 'dayjs';
 import { Theme } from '@mui/material';
-import { darkSwitch, getParams } from '../../util';
+import { getParams } from '../../util';
+import { darkSwitch } from "../../util/cssUtil";
 import type { BrokenReservation, Room } from '../../types';
 
 import Box from '@mui/material/Box';
